Inline DocumentCaptureSessionValidator
- It was only used once
- Move to Idv:: namespace, since CaptureDoc:: only had the one item

[skip changelog]
