+++ app/controllers/idv/capture_doc_controller.rb
@@ -33,7 +33,7 @@ module Idv
                 token.blank? &&
                 document_capture_session_uuid.blank?
 
-      result = CaptureDoc::ValidateDocumentCaptureSession.new(document_capture_session_uuid).call
+      result = Idv::ValidateDocumentCaptureSession.new(document_capture_session_uuid).call
 
       analytics.track_event(FLOW_STATE_MACHINE_SETTINGS[:analytics_id], result.to_h)
       process_result(result)
