--- app/services/capture_doc/validate_document_capture_session.rb
@@ -1,30 +1,0 @@
-module CaptureDoc
-  class ValidateDocumentCaptureSession
-    include ActiveModel::Model
-    include Idv::DocumentCaptureSessionValidator
-
-    def initialize(session_uuid)
-      @session_uuid = session_uuid
-    end
-
-    def call
-      @success = valid?
-
-      FormResponse.new(success: success, errors: errors, extra: extra_analytics_attributes)
-    end
-
-    private
-
-    attr_reader :success
-
-    def extra_analytics_attributes
-      {
-        for_user_id: document_capture_session&.user_id,
-        user_id: 'anonymous-uuid',
-        event: 'Document capture session validation',
-        ial2_strict: document_capture_session&.ial2_strict?,
-        sp_issuer: document_capture_session&.issuer,
-      }
-    end
-  end
-end
